Fix for genjava ignoring most packages in standalone mode
The genjava_message_artifacts tool in package genjava calls rosjava_build_tools.catkin.index_message_package_dependencies_from_local_environment()
to generate a list of all message packages and their dependencies in topological order from the list of package names given in the command line.

Especially in cases where ROS_PACKAGE_PATH lists the package paths for each individual package separately, which is the case in isolated
builds using catkin_make_isolated or catkin_tools, the relative path returned by catkin_pkg.packages.find_packages() is only `.`. In general, the
relative package path is not unique, but it is used as key of a dictionary when passed to topological_order_packages a few lines below.
As a consequence, all packages but one of each group that have the same relative package path were missing in the returned list and hence no
artifacts were generated by genjava_message_artifacts.

This patch adds a line that transforms the relative to the absolute package path, avoiding the above problem.